move codegen cli to main.ts
