+++ deno.jsonc
@@ -38,6 +38,6 @@
     "test:update": "deno task test -- -- --update",
     "mdbook:watch": "mdbook watch -o",
     "bench": "deno bench -A --no-check=remote --unstable",
-    "codegen": "deno task run codegen/codegen.ts"
+    "codegen": "deno task run main.ts"
   }
 }
